JDK-8224636: Review comments:
- use double as DashArray element

+++ modules/javafx.web/src/main/native/Source/WebCore/platform/graphics/java/GraphicsContextJava.cpp
@@ -649,7 +649,7 @@ void GraphicsContext::setLineDash(const DashArray& dashes, float dashOffset)
 
     for (size_t i = 0; i < size; i++) {
         platformContext()->rq()
-        << dashes.at(i);
+        << (float) dashes.at(i);
     }
 
     platformContext()->setLineDash(dashes, dashOffset);
